+++ api/src/main/java/ai/djl/modality/nlp/embedding/TrainableTextEmbedding.java
@@ -44,8 +44,8 @@ public class TrainableTextEmbedding extends AbstractBlock implements TextEmbeddi
 
     /** {@inheritDoc} */
     @Override
-    public int[] preprocessTextToEmbed(List<String> text) {
-        int[] result = new int[text.size()];
+    public long[] preprocessTextToEmbed(List<String> text) {
+        long[] result = new long[text.size()];
         for (int i = 0; i < text.size(); i++) {
             result[i] = trainableWordEmbedding.preprocessWordToEmbed(text.get(i));
         }
